Renamed the package to match Directory requirements.
