+++ pom.xml
@@ -3,8 +3,8 @@
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.vaadin.artur</groupId>
-    <artifactId>axa-input-text-flow</artifactId>
+    <groupId>org.vaadin.addons.mygroup</groupId>
+    <artifactId>axa-input-text</artifactId>
     <version>1.0-SNAPSHOT</version>
     <name>AXA Input Text</name>
     <description>Integration of axa-input-text for Vaadin platform</description>
