+++ src/main/java/org/vaadin/addons/mygroup/AxaInputText.java
@@ -1,4 +1,4 @@
-package org.vaadin.artur.axainputtext;
+package org.vaadin.addons.mygroup;
 
 import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.Tag;
